alloc_tracker: split the integration tests by concern
Copilot-Session: f9527c14-db4b-410b-b77f-16c19b6ecf29
